Create new Distmatrix class (with nj-method)
Previously, the Distmatrix class was in seqlib, while the phylotreelib
class had a DistTree class that then contained a Distmatrix and had
tree construction code spread over itself (DistTree) and Distmatrix.

Conceptually it is much clearer to have a Distmatrix object that knows
how to construct a tree (what is a DistTree object anyway??). It
also requires one less step (no creation of a DistTree object).

This commit contains the start of the new Distmatrix class and an
updated nj() method