Change call signature for Treefile and sub-classes
Now Treefile and all subclasses are called using this signature:
	def __init__(self, filename=None, data=None)

Either filename or data (string containing file contents) have to
be given (not both). If data is provided, then self.treefile will
be set to StringIO(data) so rest of class will work in same way
as if a file on disk had been provided.