refactor: separate handling of treefiles with transdict
Added separate handler function to NewickStringParser for dealing
with treefiles with transdict (to save testing for transdict on every
leaf - although it does not appear to actually influence performance...).

As a consequence: also had to convert handlers into instance-bound
methods (not class-bound). But that seems cleaner anyway and did not
impact performance.